requirements.txt
PyYAML>=6.0.1
cairosvg>=2.7.1
Pillow>=10.0.0 